Moved formatting functions in own module
